:recycle: Refactoring Sample
